Working Linkbar with health=100

+++ Assets/Scripts/Character/IDamageable.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public interface IDamageable
 {
 
-    void Damage(int damage);
+    void Damage(float damage);
 
     void Heal(int healing);
